requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0
python-multipart>=0.0.9
faster-whisper>=1.0
openai>=1.0
